GRsytemproject/web/requirement.txt
astroid==2.0.1
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.1.5
django-account-helper==0.1.6
django-admin==1.3.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-excel-response2==2.0.8
django-model-utils==3.1.2
django-phonenumber-field==2.2.0
django-six==1.0.4
django-suit==2.0a1
djangorestframework==3.9.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.14
oauthlib==3.0.1
phonenumbers==8.10.5
Pillow==5.4.1
pylint==2.2.2
pylint-mongoengine==0.2.2
pylint-plugin-utils==0.4
python3-openid==3.1.0
pytz==2018.9
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.2.0
screen==1.0.1
six==1.12.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
xlwt==1.3.0
